Organise classes into packages and removed unused variables
