+++ src/main/java/snipe/command/HelpCommand.java
@@ -1,3 +1,10 @@
+package snipe.command;
+
+import snipe.exception.SnipeException;
+import snipe.storage.Storage;
+import snipe.core.TaskList;
+import snipe.util.Ui;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
