+++ src/main/java/snipe/command/ListCommand.java
@@ -1,6 +1,9 @@
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+package snipe.command;
+
+import snipe.exception.SnipeException;
+import snipe.storage.Storage;
+import snipe.core.TaskList;
+import snipe.util.Ui;
 
 public class ListCommand extends Command {
     @Override
