+++ src/main/java/snipe/command/MarkCommand.java
@@ -1,3 +1,11 @@
+package snipe.command;
+
+import snipe.core.TaskList;
+import snipe.exception.SnipeException;
+import snipe.storage.Storage;
+import snipe.task.Task;
+import snipe.util.Ui;
+
 import java.io.IOException;
 
 public class MarkCommand extends Command{
