+++ src/main/java/snipe/task/ToDo.java
@@ -1,14 +1,16 @@
+package snipe.task;
+
 /**
- * The {@code ToDo} class represents a task that needs to be done,
+ * The {@code snipe.task.ToDo} class represents a task that needs to be done,
  * without any specific deadline or time frame. It is a subclass of {@link Task}.
  */
 public class ToDo extends Task {
 
     /**
-     * Constructs a new {@code ToDo} task with the specified description.
+     * Constructs a new {@code snipe.task.ToDo} task with the specified description.
      * The task type is set to {@link TaskType#TODO}.
      *
-     * @param description The description of the ToDo task.
+     * @param description The description of the snipe.task.ToDo task.
      */
     public ToDo(String description) {
         super(description, TaskType.TODO);
@@ -21,7 +23,7 @@ public class ToDo extends Task {
     }
 
     /**
-     * Returns a string representation of the ToDo task,
+     * Returns a string representation of the snipe.task.ToDo task,
      * including its type indicator and description.
      *
      * @return A string in the format: "[T][status] description".
