+++ src/main/txt/taskList.txt
@@ -1,2 +1,3 @@
-D | 0 | hi | 2019-10-15 1800
+D | 1 | hi | 2019-10-15 1800
 T | 1 | hello
+T | 0 | greet
