+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 ____________________________________________________________
 Hello! I'm
-Snipe
+snipe.core.Snipe
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
